removing wrongly placed duplicate test file
